Reworked shop + Water meter
Reworked shop using the new item implementation to be easier to use. The old implementation required making a new function for every new item button, along with creating a list of sellable items. The new implementation just requires you pass in the selling item.
Added a water meter that shows the volume of water in your bucket.