New bullet item functionality
Bullets can now be bought or grown. A bullet seed costs 1 coin but can grow 5-10 bullets. You can also buy a bullet for 1 coin. Updated Plants so that they contain a range of the amount of produce they drop upon harvest, along with a percentage chance of dropping a seed upon harvest.